Several changes large and small
Made the PEM algorithm a function. The main program now consists of  calls to the looksee, bin_errors (for errors in the model), the PEM algorithm, and the saveresults function. These have been placed in a double for loop over the frequencies (as used in Period04) and the frequency, amplitude and phase lists. The PEM algorithm is exited when any of the estimated proportions go negative. The -ve proportions are replaced with zero and the other proportion set to One. The graph labelling was also improved.